feat(server): Phase 2 move LG nodes to ./lgnodes and update imports/tests (Issue #51 Phase 2)
