+++ apps/server/__tests__/summarization.node.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'vitest';
 import { AIMessage, BaseMessage, HumanMessage, ToolMessage } from '@langchain/core/messages';
 import { ChatOpenAI } from '@langchain/openai';
-import { countTokens, shouldSummarize, summarizationNode, SummarizationNode, type ChatState, type SummarizationOptions } from '../src/nodes/summarization.node';
+import { countTokens, shouldSummarize, summarizationNode, SummarizationNode, type ChatState, type SummarizationOptions } from '../src/lgnodes/summarization.lgnode';
 
 // Lightweight mock implementing needed surface (cast to ChatOpenAI)
 const llm = {
