+++ apps/server/src/lgnodes/base.lgnode.ts
