+++ apps/server/src/lgnodes/memoryCallModel.lgnode.ts
@@ -2,7 +2,7 @@ import { BaseMessage } from '@langchain/core/messages';
 import { BaseStore, LangGraphRunnableConfig } from '@langchain/langgraph';
 import { ChatOpenAI } from '@langchain/openai';
 import { BaseTool } from '../tools/base.tool';
-import { BaseNode } from './base.node';
+import { BaseNode } from './base.lgnode';
 import { NodeOutput } from '../types';
 import { withLLMCall } from '@traceloop/node-server-sdk';
