+++ apps/server/src/lgnodes/summarization.lgnode.ts
